+++ modules/constants.js
@@ -5,6 +5,7 @@ module.exports = {
   LIBDRAGON_GIT: 'https://github.com/DragonMinded/libdragon',
   LIBDRAGON_BRANCH: 'trunk',
   LIBDRAGON_SUBMODULE: 'libdragon',
+  CONTAINER_TARGET_PATH: '/libdragon',
 
   LIBDRAGON_PROJECT_MANIFEST: '.libdragon',
   CACHED_CONTAINER_FILE: 'libdragon-docker-container',